project/llm-worker-docs/requirements.txt
torch
transformers
sentencepiece
grpcio
grpcio-tools
requests